+++ CHANGELOG.md
@@ -13,6 +13,11 @@ _Unreleased_
   `3.10.11`.  This can be accomplished by invoking `pin` with the new
   `--relaxed` flag.  #255
 
+- Adding or removing dependencies with `add` or `remove` now reformats
+  the `dependencies` array in the `pyproject.toml` file to multi-line
+  with trailing commas.  This should result in significantly better
+  diffing behavior out of the box.  #263
+
 - Default build-system and license can be specified in global config.  #244
 
 - Fixed an issue where the `init` command would not let you create